Initial Proof of concept of a HyperDeck Server written in python.
It tested with the BM HyperDeck SDK Python client.  The feature have been quickly implemented with no error checking.

My biggest issue is finding a way to get the first frame of the video to display using VLC while having the video paused.  It seem I can only get the video to display if I start playback and I need to wait a certain time to pause the video, but I found no actual specific libvlc info on when I need to pause